Factor Kore.Step.Remainder out of Kore.Step.Step
